// src/types/hr-management/resignation-management.d.ts
export interface IResignation {
    resignationFormId: number;
    employeeId: number;
    badgeId: string;
    fullName: string;
    personalEmail: string;
    workEmail: string;
    mobilePhone: string;
    companyName: string;
    dgName: string;
    dcName: string;
    projectName: string;
    positionName: string;
    grade: string;
    beforeWorkExp: number;
    currentWorkExp: number;
    totalWorkExp: number;
    paMark: number;
    personalContact: string;
    resignationTypeId: number;
    resignationTypeName: string;
    resignationStatusId: number;
    resignationStatusName: string;
    resignationStatusColor: string;
    applyDate: string;
    resignDate: string;
    cancelledDate: string;
    reasonId: number;
    reasonName: string;
    notes: string;
    isSentResignationEmail: boolean;
    isAttrition: boolean;
    isCommitmentBound: boolean;

    birthday: string;
    birthPlace: string;
    idCardNo: string;
    idCardIssuePlace: string;
    idCardIssueDate: string;
    isOfficialEmployee: boolean;
    permanentAddress: string;

    companyOwner: string;
    companyAddress: string;
    companyPhone: string;
    companyFax: string;
}

export interface IResignationEmployeeProject {
    dcName: string;
    dgName: string;
    effort: number;
    employeeProjectStatus: string;
    startDate: string;
    endDate: string;
    managerName: string;
    positionName: string;
    projectName: string;
}

export interface IResignationEmployeeCommitment {
    commitmentId: number;
    commitmentName: string;
    costFee: number;
    employeeId: number;
    fromDate: string;
    isBroken: boolean;
    notes: string;
    signedDate: string;
    toDate: string;
}

export interface IResignationEmployeeContract {
    contractId: number;
    contractTypeName: string;
    employeeId: number;
    fromDate: string;
    toDate: string;
    signOrder: number;
    isNonTechnical: boolean;
    isRenewalContract: boolean;
    renewApprovalStatusName: string;
}

export interface IResignationEmployeeOnsiteHistory {
    actualEndDate: string;
    cityName: string;
    countryName: string;
    employeeId: number;
    expectedEndDate: string;
    flightDeparture: string;
    flightReturn: string;
    onsiteCountryId: number;
    onsiteFormId: number;
    projectName: string;
    visaTypeName: string;
    customer: string;
}

export interface IResignationEmployeeTemporaryLeave {
    actualEndDate: string;
    effort: number;
    endDate: string;
    projectId: number;
    projectName: string;
    startDate: string;
    statusId: number;
    statusName: string;
    temporaryLeaveId: number;
}

export interface IResignationCancelled {
    projectName: string;
    dgName: string;
    dcName: string;
    applyDate: string;
    cancelledDate: string;
    reasonName: string;
}

export interface IResignationMoreInformation {
    employeeCommitments: IResignationEmployeeCommitment[];
    employeeContracts: IResignationEmployeeContract[];
    employeeOnsiteHistories: IResignationEmployeeOnsiteHistory[];
    employeeProjects: IResignationEmployeeProject[];
    employeeTemporaryLeaves: IResignationEmployeeTemporaryLeave[];
    resignationFormCancelled: IResignationCancelled[];
}